// frontend/src/interfaces/data/data.ts
export interface IStarting {
    name: string;
    link: string;
};

export interface ISkill {
    name: string;
    icon: string;
};

export interface IHubProject {
    number?: number;
    name: string;
    image: string;
    subject: string;
    skills: ISkill[];
    difficulty: number;
    startings?: IStarting[];
};